+++ chaincode/ledger/dbal_algo.go
@@ -62,16 +62,6 @@ func (db *DB) QueryAlgos(p *common.Pagination, filter *asset.AlgoQueryFilter) ([
 	if filter != nil {
 		assetFilter := map[string]interface{}{}
 
-		if len(filter.Categories) > 0 {
-			categories := make([]string, len(filter.Categories))
-			for i, c := range filter.Categories {
-				categories[i] = c.String()
-			}
-			assetFilter["category"] = map[string]interface{}{
-				"$in": categories,
-			}
-		}
-
 		if filter.ComputePlanKey != "" {
 			algoKeys, err := db.getComputePlanAlgoKeys(filter.ComputePlanKey)
 			if err != nil {