+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Category filter from QueryAlgos rpc (#836)
+
 ## [0.20.0] - 2022-07-25
 
 ### Added